pipelines/fasterwhisper/requirements.txt
# Transcription
faster-whisper==1.0.3

# Diarization
pyannote.audio==3.1.1

# PyTorch (compatible version without weights_only issues)
torch==2.5.1
torchaudio==2.5.1

# Core dependencies
numpy>=1.24.0,<2.0.0
requests>=2.31.0
tqdm>=4.66.0
huggingface-hub==0.19.4
